Refactor stacktracelink
